refactor: better path structure
