+++ lib/mutations/event-types/update-event-type.ts
@@ -1,4 +1,4 @@
-import * as fetch from "@lib/core/api/fetch-wrapper";
+import * as fetch from "@lib/core/http/fetch-wrapper";
 import { EventTypeInput } from "@lib/types/event-type";
 import { EventType } from "@prisma/client";
