+++ pages/sandbox/preview-error-page.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from "next";
 import { ErrorPage } from "@components/error/error-page";
 import React from "react";
-import { HttpError } from "@lib/core/error/http";
+import { HttpError } from "@lib/core/http/error";
 
 const PreviewErrorPage: NextPage = () => {
   const statusCode = 403;
